discoveryf4: rename DiscoveryF4 target to FlyingF4
The FlyingF4 target is expected to include all interface
definitions for enough external ports to allow the board to
fly a quad.

The DiscoveryF4 target will be a future target that is
strictly made up of components on the eval board and will
be useful as a very basic test platform only, not for flying.

Commands used to do the bulk of the renaming for this target:
  git mv flight/targets/{Discovery,Flying}F4
  git mv flight/targets/Bootloaders/{Discovery,Flying}F4/
  git mv flight/targets/board_hw_defs/{discovery,flying}f4/
  git mv make/boards/{discovery,flying}f4/
  git mv flight/PiOS/Boards/STM32F4xx_{Discovery,Flying}F4.h
  git mv flight/targets/FlyingF4/System/{discovery,flying}f4.c
  git mv hardware/Production/{Discovery,Flying}F4/
  git mv flight/PiOS/STM32F4xx/Libraries/CMSIS2/Device/ST/STM32F4xx/Source/{discovery,flying}f4/

Then, a few edits to fix up references to the board name.