usb: Use SOF interrupt to determine that USB is connected
This allows for a fast (250ms) detection of a USB cable
being connected on platforms that don't support a VSENSE
pin.

This is particularly useful on the FlyingF3 platform.

The 250ms was determined empirically on a Linux system.
This may need to be tweaked once we have results from
other OSs.