flyingf3: increase BL delay when waiting for USB cable
The previous delay of 250ms was too aggressive in
certain circumstances.  If the USB plug was inserted
slowly, the BL would power up ahead of USB D-/+ being
properly seated.  Also, when connected via a hub,
there was sufficient lag while the hub was enumerated
that the BL would not notice the USB cable was connected.