api: let all firmware builds see the API shared with the GCS
This API allows some internal structures to be shared between
the GCS and the flight code.

This should ONLY be used when you need to establish some kind
of additional structure inside of the data fields inside of a
UAVO.  This should NOT be used to avoid using a UAVO for any
reason.