i2c vm: move UAVO init into module initialize function
All UAVOs must be initialized in the module init() routine
and *NOT* in the start() routine or they may not be noticed
by the Telemetry module.

The telemetry module scans the list of UAVOs at the beginning
of its start() routine.  Any UAVOs initialized after this scan
will not be noticed by the Telemetry module, and as such will
also not have any update events registered against them.

This will lead to the UAVOs never being updated via telemetry.
This affects periodic as well as onchange modes.