Packaging: a package_matlab target
Previously the way matlab was being packaged it was required for
packaging but wasn't actually done before the scripts which make
the platform specific bundles was run.  The code still should be
refactored so that there is a clear set of things built, then the
packaging step.  However, this fixes it for now.

Also make sure the matlab packaging depends on uavobjects