ManualControl: Position hold should set path desired to HOLDPOSITION
Previously was using FLYENDPOINT mode.  Both should be acceptable but the
vtol code expects for FlightStatus = PositionHold the PathDesired.Mode should
also indicate the same.  I'm conflicted if this is good since it's essentially
redundant modes, but giving more rich information to the follower might
prove useful in the future (e.g. the way you hodl a position might be different
than flying to an endpoint, even if the current code doesn't treat it as such).