Android PFD and SystemAlarms: Ensure that objectUpdated is not called for a null object
Both SystemAlarms and PFD (fragments) are getting a connected signal and then not finding
objects that should exist (SystemAlarms and AttitudeActual respectively).  This is probably
when connected to an unknown UAV type and should be handled gracefully like now.
However, it would be good to confirm this doesn't happen when connected to a known
UAVO set.

fixes #203