Android: Better handling of broken telemetry channels
Previously because the error occurred within a runnable on the master
telemetry thread I was not properly propagating the error to the higher
level of control.  Now it causes the master thread looper to quit
which then calls telemetry disconnect and shuts everything down
properly.  This is tested with network telelmetry and allows reconnecting
after the remove network server is shut down.

Fixes #236