DSP Lib: Order of operations undefined
GCS (and clang) seem to not like statements like

uint32_t *pT1;
*pT1++ = a;

Because it is unclear if you are incrementing the pointer and then
assigning or assigning and then incrementing.  I'm also not sure
and there is a chance I picked the wrong ordering.  Because of the
fact the math is meant to iterate over the columns and rows I
believe in intends an assignment and then a pointer increment and
coded this explicitly.