Ground Vehicle: Add a car motor type
Because car escs interpret (1000,1500) us as negative the normal
motor output type is inappropriate for cars as forcing the output
to minimum will knock them into reverse.  Using the servo output
though provided no protection for arming.

This adds a new output type "CARMOTOR" which when disarmed goes
to neutral instead of minimum.  In failsafe it also goes to
neutral.

This is pretty much a "future proof" bug and doesn't actually
provide reverse because the throttle curve itself only outputs
positive values.