INS: Simplify initialization and remove redundant code
Previously there were two branches of code for indoor and outdoor mode
in the INS that were very similar and then some complicated logic to
try and run the prediction forward without correction for a while
when starting the INS.  This seemed to result in more poor initialization.

Also reduced the (hardcoded) indoor variance to 0.1.