Qt5: On mac have the default style "fusion" and use it for the tab widget
For some reason the tab widget with Manhattan looks really bad and
critically the dark color wraps around where the connection buttons
are and shows up at the top.  In the long run maybe we can drop
manhattan entirely, but for now doing so makes the widget selector
dropdowns huge.