Qt5: Add qmake bin directory to path
make V=1 gcs fails on windows if the only thing in the
path is mingw32-make because it cannot find the path to
the qmake dll's.  This adds it to the environmental path
in the build environment.

In this configuration, running GCS will still fail unless
you make release, which copies all the dll's to the appropriate
location.  Possibly there should be another solution that a
more windows-familiar person would know.