Camera Tracking: Process tablet updates with rest of information
Only enable tablet callback when needed

Previously converting the tablet position into NED location was
run within a callback which isn't ideal, and also would run as
often as it updated - regardless of how frequently it was used.

Now it is within the camera thread (unfortunately still a callback)
but only when needed and at a limited frequency.