UAVTalk: Fix crash when requesting update of unknown object
UAVTalk would check whether an unknown object was single instance
even if the object type was unknown, which caused a segfault.  This
value is only used to validate the payload length (possibly an
unnecessary check) so now that check is not applied if the object
is unknown.

The response to an OBJ_REQ for an unknown object is simply to return
a NAK packet so this should keep multiple instance objects ACK'ing
properly while solving the segfault for unknown objects.

Any non OBJ_REQ message should not even reach this far as they should
error our with a bad object type.

Fixes #80
Fixes #358