Stabilization: Check that CameraDesired exists before using
When using POI mode it is possible that the CameraStab module is
not running so we need to check this handle to avoid a segfault.
Set an error in this case, but ideally this should be handled in
SanityCheck.  Currently it is for normal modes but not for
navigation.  However the later case is less critical since if you
can use this POI heading mode it is probably safe to leave alone.