f4 usb: always init HID and CDC devices
This is required so that we can always respond to HID and
CDC requests via the control endpoint (ep0).  The F1/F3
implementation gets away with disabling these interfaces
because all of the ep0 handling is in the usbhook code
rather than (properly) split out into class-specific
drivers like it is done on the F4 implementation.