GCS Log Replay: Fix replay of android log files
The android log files use the android system time which means it
can start at very large numbers, whereas GCS logs start at zero.
To fix this the time of the first timestamp is subtracted out.

Also the times were signed which is inappropriate, so they are
all converted to quint32.

As a side note the way this code is written, it is _extremely_
brittle to ill-formatted logfiles because it doesn't use the
UAVTalk parser or try and maintain sync.  I also removed a
redundant set of checks on the order of timestamps as this is
done during loading and also didn't respond well to the large
timestamps from android.