GCS Log Replay: Fix replay of android log files
The android log files use the android system time which means it
can start at very large numbers, whereas GCS logs start at zero.
To fix this the time of the first timestamp is subtracted out.

The times were signed which is inappropriate, so they are all converted
to quint32. This avoids any integer overflows when comparing with the
Android timestamp.

Also removed a redundant set of checks on the order of timestamps as this
is done during loading and also didn't respond well to the large timestamps
from android.

From peabody124.