Sparky: Probe for MPU9150 before using MPU6050
The MPU9150 is essentially an MPU6050 with an additional mag
inside so probe for it before the MPU6050.  This might not be
the way to go in the long run - we might want a separate driver
for the mag and MPU6050 and run them independently.  However a
little code duplication means we can avoid situations of bus
contention more easily.