Optimized transmitter calibration wizard metadata rates.
Also solved bug by which update rates were incorrectly reset during
the wizard. In this case, there was a metadata reset followed immediately
by a set. This reset/set operation occurs too quickly and the telemetry
system only sends the first operation to the board. The result was that the
final steps of the wizard were very slow.