Revert "Revert "Added qml files to qmake project files.""
This reverts commit 7c1fe890d925259e95fb2d524e168e13bdb5ad47.

The reverting author was mistaken, as the reasoning for the revert
was incorrect. The files are already part of the plugin so including
them in the OTHER_FILES is appropriate. Refer to Qt Documentation
for how to use OTHER_FILES.

They do need to be moved to another location, as per #460.