Revert "Revert "Revert "Added qml files to qmake project files."""
Per discussion on IRC the consensus is that including files from the
share directory in a plugin .pro is bad design.  By definition these
files are not part of the plugin.  In this case, the QML PFD is designed
to present a walled garden for graphical designers who can create
multiple sets of PFD artworks.

This means #460 should probably be marked as wont fix too, unless
we want to lose the ability to select between multiple PFD UIs
which would be a shame.

This reverts commit 4def50c6adfd3ccd17d9e1f3cfd41da76ebd174d.