Path Saving: use explicit indication for end of path
Previously once a waypoint could not be loaded that was taken as
the end of the path.  However, that doesn't allow distinguishing
a lost waypoint from actually getting to the end.

Now the WAYPOINT_MODE_STOP is used to indicate the end when saving
and loading.  Only will indicate COMPLETED for load when a STOP
is read from flash.