Android Audio: prevent GPS status messages backing up
When the GPS is near threshold a lot of alarm togglings can occur
which saturates the audio and blocks up other messages.  This only
lets one GPS alarm and one GPS clear message exist in the queue.
This could occasionally result in the last GPS message being stale
but there is no clear way to remove particular messages from the
audio queue.

Later we might want to do the queueing explicitly in the user space
to prioritize messages.