Path Filleting: present dialog to select radius
Previous filleting radius was hardcoded to 5 meter but
now a dialog will be presented to allow changing this.

Unfortunately because the dialog should have a handle to
the parent widget this requires adding an optional parameter
to the processPath call which seems unclean.  Really the UI
handling should be separated.