StateEstimation: module that utilizes the new filter architecture
This supports SE(3) filters and generic filters currently.  In the
case of a generic, it is expected to fetch all of its own data.

An important TODO for this (and for generics) is the ability to
disconnect queues not in use.  For example since the core code
provides an infrastructure for SE(3) filters that should only be
enabled in the case of running and SE(3) filter and should be
disabled otherwise.  At this point it will probably make sense
to move process_filter_s3 into its own file within the module
(since this is still not really filter code) and add two functions

init_filter_class_infrastructure_s3
deinit_filter_class_infrastructure_s3