SetupWizard: use IBoardType::InputType enum
Previously the SetupWizard had its own enum for the input type.
This then got mapped to the IBoardType value which was then
mapped to the specific hardware (in the board specific plugin).

It doesn't make a lot of sense to have basically two equivalent
enums, so now the setup wizard uses the board plugin enum.  The
opposite direction wouldn't make sense as we don't want to induce
a dependency of the setup wizard on the board plugins.