flash partitions: add flash partition layer between filesystem and chip
This new partition abstraction improves on the previous division
between the filesystems and the chip drivers.

The partition API allows the users of flash to be unaware of the details
of the sector layouts and max read/write page sizes of the underlying
devices.  Filesystems are now mounted on partitions by the ID of the
partition rather than using specific drivers and offsets within a chip.