SetupWizard: rework abstraction to support other wizards
Right now all the wizard pages are given a SetupWizard in their
constructor which precludes using this architecture for other
wizards.  Almost all of the time they don't really need to use
the parent wizard and only access it because either
   1) it is a QWizard for things like calling close
   2) it is a VehicleConfigurationSource and they want to work
      with that

The VehicleConfigurationSource should probably be abstracted out
of the actual wizard so the wizards _have_ a VehicleConfigurationSource
and not so that they _are_ a VehicleConfigurationSource.  This is
generally a better design pattern when possible.  It also means we
can easily drop in different configuration sources with a standard
API which can come from the board plugin system.

For now these changes start that by making the AbstractWizardPage
initialize with any QWizard.  For backward compatibility getWizard()
dynamically casts back to a SetupWizard* so all instances of using
this should be changed to getQWizard() when possible.