NavigationWizard: do not inherit from VehicleConfigurationSource
The VehicleConfigurationSource has a bunch of pure virtual functions
that make sense for a general SetupWizard but not for specific aspects
like navigation.  So the navigation wizard does not inherit from it
anymore.

The SetupWizard functions using this should be moved _into_ the
VehicleConfigurationSource and it should forward calls (or let
the pages directly access its VehicleConfigurationSource object).