AbstractWizard: add getQWizard getObjectManager methods
The getQWizard is based on the previous commit comments.  Make a method
to get the minimal interface required to establish a better abstraction
from wizard methods and configuration methods.

The getObjectManager helps for direct access to objects.  For consistency
with the current implementation it should be used judiciously because all
the code to actually write to the flight controller is consolidated in the
VehicleConfigurationHelper class.  So really this should only be used for
getting settings.