watchdog: clear wdg backup register during init
Using the PIOS_WDG_REGISTER slot in the backup registers is
a useful debug tool but it is important to ensure that this
state cannot influence the watchdog behaviour on future reboots.

Specifically, if a future boot removes one of the modules that
was previously watched by the watchdog, we don't want that
removal to endlessly trigger a watchdog.

F1 variant was actually clearing the flag, but was actually
doing it *before* caching the value for debug purposes.  F3/F4
variants were carrying state across a reboot which is bad for
the reasons listed above.