FlyingF4: enable storing waypoints in external flash
A flag PIOS_INCLUDE_EXTERNAL_FLASH_WAYPOINTS enables storing the waypoints
in an external flash chip, which allows for many more than internal storage.
This also preserves the code for using the external flash chip.

Alternatively waypoints could be stored internally on the F4 but since there
are very few FlyingF4 users this is somewhat arbitrary.