Issue #720 update UAVTalk protocol to send NACKs when receiving unknown ACK'ed objects
This update introduces a backward-compatible update to the UAVTalk protocol: when receiving
an object for which an ACK is expected is return, and the receiver does not know the object
(ObjectID unknown), then instead of dropping the object, send a NACK with the ObjecID.

The most immediate side effect is that this avoids a situation where the GCS tries to
send a settings object that is unknown to a board, and goes through multiple timeouts. Now,
when sending a settings object that the board does not know, the transaction will fail right
away.