ground/maplib Library rebranding third pass
Added TauLabs as the author on all lib files.
Original author was mantained.

The original library author was contacted and he wishes that the
branding be related to the project which actively keeps the library up
to date.
Since the original project doesnt seem interested we are doing this
ourselfs and as such rebranding it.