F3 I2C: fix for first transaction
Do the same thing as F1/F4 with regard to the sem_ready semaphore
which is to make sure the driver grabs it for before the first
transaction and then wait for it to be available again.  This
is an alternative to the fix suggested in 755 but consistent
with the other drivers.