Virtual flybar: use PID controller
Previously the virtual flybar mode was just a PI controller
with the P always on the gyro (setpoint of zero) and the I
applied to the virtual flybar angle.  The stick position
suppresses the controller output and adds a command signal.

Now instead of the simple P controller component the PID
controller is used but the ILimit is zero (unused) and the
PD component is used.  This is still suppressed the same way
but allows tuning a Kd parameter to get a higher gain with
neutral stick.