heap: reserve early init heap space in linker map for f1
The F1 targets use the tail of the heap as the stack for the
init code and then after init, this RAM is added onto the heap.

This init dance was relying on a statically allocated
standard heap.  The new heap allocator doesn't statically
allocate it but rather relies on the _sheap and _eheap
symbols being correct.

This change forces the linker to reserve some portion of the
heap to be available during init.