f4 usb: make board-specific copies of usb_conf.h
The revolution and revomini f4 targets need a different
setup for usb_conf.h.  This is (currently) due to the
revo/revomini boards needing to disable the
   VBUS_SENSING_ENABLED
configuration since they don't connect VBUS to the dedicated
USB peripheral pins.

All boards now carry their own independent copy of this file
so that the conflicting requirements can be satisfied.