bl unified: use cable connected instead of check available
PIOS_USB_CheckAvailable() checks for both cable connected
as well as whether the device is fully enumerated.  This
more stringent check takes a more variable amount of time
on different computers.

It is much more appropriate to simply check whether the
USB cable is connected to determine whether we should wait
further in the bootloader.