makefile: require ARM toolchain to be installed under ./tools
The practice of searching for the toolchain in the user's
PATH when we can't find the toolchain under ./tools is
problematic.

If the user happens to have an old or incompatible ARM gcc
in their PATH and we change the toolchain version, we will
silently run the one in their PATH and possibly produce
a broken executable.

We now fail the build explicitly if we can't find the correct
toolchain version under ./tools.