InputWizard: cancel out when board is disconnected
If a user plugs in a board and then unplugs it during the middle
of calibration their original settings (including arming) will be
restored.  However, if they then start moving the sticks to test
the range they will end up arming the system which is dangerous
(although props should have been removed to do this step).