SanityCheck: prevent unsafe forms of arming
Arming while in autonomous modes (e.g. RTH, PH, AH) can
allow unexpected behavior and is certainly an untested
path. This patch creates an error condition when in those
modes and disarmed or arming to prevent arming. In this
case one should switch to a traditional manual mode, arm,
and then engage the AP functionality.