Horizon: connect inputs to the PID controllers
This mode uses the raw stick input to smoothly transition
from an outer attitude controller setting the desired rate
and the stick input. This allows switching from an attitude
mode to a rate mode continuously.

This preserves the exponential behavior of rate mode if it
is engaged so will keep both the attitude component and the
rate component trimmed the same as the basic modes.

This mode is modeled off the multiwii controller code, specifically
from the baseflight fork in the pidMultiWii. One difference is
suppressing the outer loop integral component. Also in that code
the "rate" mode doesn't appear to use an acutal proprotional
component despite the name and just pushes the stick input
straight to the output and uses an integral controller on the
outside (more like our virtual flybar mode at the extremes).