Horizon: blend amount determined by greatest stick
Previously the blend between attitude and rate was done
independently for each axis. However, that meant when
performing a pitch flip, the roll axis was on attitude
mode and would try and keep level. Once going inverted
the controller would do a roll to attempt to level out,
which is undesirable.