VTOLPath: use landing rate setting
I missed that this was in the UAVO so using that setting now.

Of note - there is a serious concern that if the EKF vertical estimate
is far enough off this could result in going up instead of down. Since
there is no second outer loop tracking position it would not be caught
or corrected for.

The correct fix for this is to make sure the EKF is robust to errors in
the z-accel bias within a reasonable range.