VTOLPath: enable the PathPlanner functionality
This isn't the most elegant code. When the PathPlanner is enabled
as a flight mode, the follower FSM enters a state that constantly
checks for the latest value of PathDesired and tries to follow it.

The PathPlanner entirely takes care of setting that PathDesired
value, which means that the PathPlanner cannot utilize any of the
sequencing abilities of the PathFollower. In the future it might
be good to add a "Goal" field to the PathDesired which is then
either interpreted as _input_ to the Follower when in path following
mode or set when the PathFollower is being configured by itself (e.g.
when a flight mode is triggering it to run).

So basically there should be a better distinction between exogenous
input (path following) and internally generated input (flight modes
that trigger follower).