Android: clean up fragment lifecycle
Instead of trying ot make the base activity manage all
the message subscriptions for a fragment, it makes much
more sense for the  base fragment to do this directly.

This also creates two ObjectUpdate methods. One for when
the UI is available and another when it is not. This means
that when an fragment is added to the backstack it can
continue to acquire data or be informed of changes - and
thus when restored have the latest data.