Android: more work on the fragment lifecycle
This now seems to work pretty well. When they are destroyed
the callbacks are all removed and the objMngr disconnected
event is simulated (so the individual fragments can clean
anything up there).

This also means that while in the Stopped state updates
will continue to accumulate. If you only want to receive
updates while running override the objectUpdatedUI method
instead.