gcs input widget: fix flightmode position calculation
Range was effectively being computed/split across 0..N-2 and
would only ever show N-1 (ie. the max flight mode) when
valueScaled was exactly 1.0f.  Even the slightest jitter on the
flightmode channel would result in the GUI drawing the flight mode
as N-2.

The code in the GCS now matches the code in the flight firmware.