Python: logparse.py now deals properly with Overo logs
The -t flag tells the parser to expect the timestamps to be
in the UAVTalk packet instead of the hacky format that prepends
them. In addition, when the git hash is passed on the command
line the header file is no longer looked for.

A small fix to the uavtalk.py is in here related to parsing the
packets with timestamps in, because the get_size_of_data method
for uavo.py now returns the size of the DATA and does not account
for the size of the timestamp.