INS: allow a default mag value for indoor mode
Since we do not perform navigation in this mode, requiring
a GPS to get the HomeLocation or manually setting it is not
necessary. Also, forcing people to set it in the map is
bad because that saves by default - and then they might
get in the field with a very incorrect home location.