Attitude: add NONE navigation filter
We previously tried to use a RAW mode for the GPS to pass directly
for navigation, but this could also be confounded with situations
when you didn't want GPS. This created an ambiguous situation and
prevented us performing sufficient safety checks.

Now the user either expressly selects "NONE" for no navigation, and
if they attempt it there is an error.

If they select RAW and there is no valid GPS signal, it will also
set an error.