Autotune: check in new algorithm
This now makes small controlled twitches on all axes
and then fits the parameters of the transfer function

                beta
             ---------
            s (1 + s * tau)

to the gyros and control outputs, which captures the
delay of the ESC in tau and the strength of the motors
in beta.  It also captures the bias in the output, but
that is not used in tuning.