Ground: updated autotune equations
This uses the tau, beta, and noise values measured from the system
identification and computes the Kp, Ki, and Kd parameters for the
rate loop.

There are a number of constrains that went into the derivation of
these equations. It allows setting the pole (a) of the single
order root and the undamped frequenchy and damping coefficient
of the second order component.

That least one free component which is determined by limiting the
amount of noise of the transform at high frequencies.