matlab: system ident only use non-zero covariance terms
Store the covariance matrix in a linear array and generate
C code that has the indexes into the full array replaced with
those mapping into the spare array of non-zero elements. This
also allows only working with the elements in the upper
triangular component.

This reduces the covariance storage from 169 floats to 43, which
is twice those savings since the array must be duplicated.