Ground: optimize the Ki rate terms too
Uses new equations that calculates the integral component
in a more principled manner. This creates a fourth order
characteristic equation and tries to place all the poles
at the same location to get the fastest response.

The main limit on speed is what the derivative term can
be, which influences the high frequency noise sensitivity.