UAVO Browser: Reset proxy model when regenerating UAVO tree
This resolves an issue by which the tree would behave poorly when changing
a view mode, e.g. enabling/disabling scientific notation or category view.
For a reason I do not yet understand the proxy model has to be deleted and
then recreated in order to capture the reset in the underlying UAVO tree
model.

Remove the Q_ASSERTs that were being used for debugging purposes.