R70008SB: correct implementation of end byte detection
According to Rusty's data sheet these are the possible
end byte values

 *    00000100
 *    00010100
 *    00100100
 *    00110100

Previously doing a modulo would only detect the second
end byte, which makes the code work but drop 3/4 of the
frames.

The correct implementation is to mask out the 0x30 bits
and check for the constant 0x04. It would also be possible
to have a more robust check for lost packets by looking at
the sequence since Rusty's original diagnostic found it was
deterministic. However, since we don't currently have a
consistent way to determine RSSI from the driver this isn't
done.

This code can be verified by the following python code:
for a in [int('0b00000100',2), int('0b00010100',2), int('0b00100100',2), int('0b00110100',2)]:
   ....:     print `(a % 0xCF, a & 0xCF)`
   ....:
(4, 4)
(20, 4)
(36, 4)
(52, 4)

Showing a mask produces a constant value.