RFM22B Rcvr: variant of PIOS_GCSRCVR for RFM22B module
This is from pios_oplinkrcvr which was from pios_gcsrcvr. Renaming
and reformatting because it applies for any modem board with an
RFM22B module running our code.

The driver is designed to run on the _flight controller_ side and
catches updates from a UAVO and presents that on a pios_rcvr
interface. Later, it might make more sense to have a special packet
type that the RFM22B uses to represent PPM information rather that
requiring it to stay synced to a UAVTalk stream.