TauLink: enable PPM input and relaying
This also changes the option a bit. Simplifying things by only
considering the case that the modem sends PPM. Whenever the PPM
port is enabled it will attempt to relay any detected stream via
the RFM22BReceiver object.

The modem will be in a coordinator mode (attempt to establish
connection) when either in TelemCoord or PPM. However, even if
not the coordinator it will still send PPM if that port is enabled.