GCS: confirm failsafe is working during input wizard
Some receivers have a failsafe mode that do things like hold
the throttle at a constant setting. In this situation, the flight
controller cannot tell that the pilot is no longer in control.

This is an extremely unsafe configuration and should not be
facilitated by the wizard. This adds a step that requires the
user to verify that failsafe detection is working reliably.