Stabilization: implement TPA based on throttle setting
Calculate the desired PID suppression based on throttle
settings. This is similar to an algorithm used by MultiWii
and empirically works well. It creates a piecewise linear
suppression of PIDs versus throttle.

i.e. the Kp will look like this:

     -------------         <-- rateKp
                  \
                   \
                    \      <-- rateKp * (1 - attenuation / 100.0)
                 ^
                 |
                 threshold

For safety the maximum allow attenuation the firmware will pay
attention to is 90%